[memory64] Avoid using i64/BigInt for syscall return values. NFC
The return values of syscalls are always errno value so there is no
need to using long/i64 for them.   This simplifies the generated JS
code a little and reduces the amount if i64 usage and the JS/wasm
boundary.

+++ system/lib/libc/emscripten_mmap.c
@@ -52,7 +52,7 @@ static struct map* find_mapping(long addr, struct map** prev) {
   return map;
 }
 
-long __syscall_munmap(long addr, long length) {
+int __syscall_munmap(long addr, long length) {
   LOCK(lock);
   struct map* prev = NULL;
   struct map* map = find_mapping(addr, &prev);
@@ -95,7 +95,7 @@ long __syscall_munmap(long addr, long length) {
   return 0;
 }
 
-long __syscall_msync(long addr, long len, long flags) {
+int __syscall_msync(long addr, long len, long flags) {
   LOCK(lock);
   struct map* map = find_mapping(addr, NULL);
   UNLOCK(lock);
@@ -108,7 +108,7 @@ long __syscall_msync(long addr, long len, long flags) {
   return _msync_js(addr, len, map->flags, map->fd);
 }
 
-long __syscall_mmap2(long addr, long len, long prot, long flags, long fd, long off) {
+int __syscall_mmap2(long addr, long len, long prot, long flags, long fd, long off) {
   // addr argument must be page aligned if MAP_FIXED flag is set.
   if (flags & MAP_FIXED && (addr % WASM_PAGE_SIZE) != 0) {
     return -EINVAL;
